requirements.txt
Django==2.2.7
django-crispy-forms==1.8.0
django-markdown-deux==1.0.5
django-pagedown==2.0.3
markdown2==2.3.8
pytz==2019.3
sqlparse==0.3.0
